best effort code block html escaping
this text block isn't long enough to explain this tiny change.

there are many ways to render MD but not HTML in user supplied text - however, most sanitizers remove instead of escaping HTML in user inputs. I tried a few approaches before settling on this best-effort approach:

1. Replacing all <>s with their HTML escape, then providing a custom MD renderer to transform those escapes back inside code blocks (not by detecting the code blocks via regex, like this final approach). This caused angular to detect the valid HTML inside the code blocks and remove it in it's sanitization step.

2. Using marked's sanitize option. This worked more or less as expected, but the option was deprecated and going to be removed (https://github.com/markedjs/marked/issues/1232).

3. Using an HTML sanitizer. I tried both DOMPurify and html-sanitizer (only the latter of which had an escape instead of remove option) to sanitize user input both before and after markdown rendering. This encountered the same issue as approach 1.

In the end, I fell back to a best-effort regex to detect transformed <>s in code blocks and transform them back. If something manages to escape this step, Angular's built in sanitizer will catch and remove it, but it's unlikely anyone will encounter an edge case that does unless they are intentionally trying to break out.